Added UserStart() in Voucher.cs

+++ Project/DataAccess/VoucherAccess.cs
@@ -34,7 +34,7 @@ public static class VoucherAccess
         return _connection.Query<VoucherModel>(sql).ToList();
     }
 
-    public static List<VoucherModel> GetVouchersByUserId(int userId)
+    public static List<VoucherModel> GetVouchersByUserId(long userId)
     {
         string sql = $"SELECT id AS Id, code AS Code, description AS Description, amount AS Amount, type AS Type, user_id AS UserId FROM {Table} WHERE user_id = @UserId";
         return _connection.Query<VoucherModel>(sql, new { UserId = userId }).ToList();
